Change style! macro syntax
Change style! macro so that the id takes a string literal or a
variable that holds a string. This makes it easier to work w/ rather
than using an `Ident`. This is replaced w/ an `Expr` which evalates
into a String (via a call to .to_string()) in codegen.rs.