Keypress explicity matches on KeyEventKind::Press
Crossterm 0.25.0 added a new field in KeyEvent, called `kind`.
https://github.com/crossterm-rs/crossterm/blob/10d1dc246dcd708b4902d53a542f732cba32ce99/src/event.rs#L645

Currently in terminals that do NOT support kitty keyboard protocol
(https://sw.kovidgoyal.net/kitty/keyboard-protocol/), which is just
regular terminals, the `kind` is always `Press`. This is made
explicity in the code in this commit.

Rename Key::Enhanced to Key::KittyKeyboardProtocol