Fix bug where caret can end up in middle of grapheme cluster
In EditorBuffer, when pressing up/down it was possible to move the
caret into an invalid position, right in the middle of a grapheme
cluster segment. This is remedied by the `validate_cursor_position`
function in line_buffer.rs, which ensures that the caret will be
checked for validity and reset if needed after up/down is applied.