Change semantics of EditorBuffer empty state
EditorBuffer used to start w/ an empty vec![] of lines.
This change automatically inserts a "" into this vec![]
so it always starts w/ a single line.

This change makes more sense in the context of an editor,
where it is strange if the editor component starts w/ a new
buffer and it paints that the buffer is empty. Then the
user has to type something for it to "get it started".