Add new RenderOps to make it easy to print text w/ auto padding
With https://github.com/r3bl-org/r3bl_rs_utils/issues/54 it became
apparent that it is necessary for each RenderOp that paints text to
take care of it's entire display area. The easy solution was to add
empty spaces to fill in that area after painting (to remove any
ghosting from a previous render).

RenderOps::PrintTextWithAttributesAndPadding makes this trivial.