Change Style id from String to u8
There was never any good reason for Style to have a String id. Styles
are calcuated and copied all over the place. Due to it using String,
it was not possible to implement the Copy trait.

By using IntEnum and u8 for style id, this solves the problem & Copy
trait can now be implemented.

Removed all the needless clone() calls throughout the codebase &
updated all the tests.

This should improve rendering performance, since there are far fewer
String clones that are being done for every re-render. Have not
verified this by running the flamegraph profiler yet.