Rewrite Style to add italic attribute
For some reason Style was missing the `italic` attribute. In order
to add it quite a few changes are made along the way in the name of
code cleanup and removing dependencies that aren't needed.

One big change is getting rid of bitmask for Style, since this was
just more headache than it was ever worth. The current implementation
is much simpler and easier to understand and maintain.

This 2nd and 3rd order effect of this change how render ops are
generated and a whole host of other infrastructure. A lot of the
code is changed to be idiomatic Rust as well, since some of this
is about a year old at this point.

Along the way lots of documentation & formatting are done.