[core] remove `reqwest` and license exception `encoding-rs`
The commit removes the use of `reqwest` crate from the core code. Since
the `encoding-rs` was dependent on the reqwest code and we were expecting
an license exception for `enoding-rs`. But now we don't have reqwest,
thus there is no encoding-rs in dependencies and can be safely removed
from license exception.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>