[tui] feature : Copy Selected text in TUI
The commit comprises of changes required to implement the feature where
we can copy the text to the clipboard from TUI. Here, We set in a logic
which takes whole text and the `SelectionMap`. We then take the RowIndex
in order and then based on it extract the selection caret values from all the
lines(rows). Then we pass this to clipboard crate API to copy with our
system clipboard.

Another minor change which have been made is for `get_as_string()`
method for SelectionMap is now using join delimeter as a newline instead
of comma.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>