[tui] Paste the text from clipboard in editor component
The commit comprises of changes required to implement the feature where
we can paste the text which was copied in the clipboard. Here we use the
internal function which inserts a chunk of string at a given caret
position by extraction the required contents from the clipboard API from
the crate.

Also, the copy_to_clipboard module has been renamed to clipboard_context
so that it can be used as a generic module which integrates all the
future features under one module and makes the code easily
understandable and managable.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>