[tui] Editor component - caching content to avoid rendering if unmodified
Cache the rendered output when there is no modification in the content
or the change in scroll offset and window size. This will avoid
rendering whole content for each movement of caret or even in selection.
Caching the content should improve the performance of the editor buffer
since we avoid rendering on each action.

Signed-off-by: Harshil Jani <harshiljani2002@gmail.com>