[core, tui] Refactor Cargo.toml and remove syntect from core, change all ambiguous names to be explicit
This is an attempt to isolate the use of syntect to just 1 crate: r3bl_tui.
It has been removed from the core crate. However, it is not possible to remove
syntect from the lolcat / color_wheel modules as they are intrinsically tied
together. Basically, they need to be able to render output, and for that they
need to be in the r3bl_tui crate which has a dependency on syntect and this
can't be removed.

safemem is no longer maintained, you can see the following output from running
cargo deny check advisories.

180 │ safemem 0.3.3 registry+https://github.com/rust-lang/crates.io-index
    │ ------------------------------------------------------------------- unmaintained advisory detected
    │
    = ID: RUSTSEC-2023-0081
    = Advisory: https://rustsec.org/advisories/RUSTSEC-2023-0081
    = The latest crates.io release was in 2019. The repository has been archived by the author.
    = Announcement: https://github.com/abonander/safemem
    = Solution: No safe upgrade is available!
    = safemem v0.3.3
      └── line-wrap v0.1.1
          └── plist v1.6.0
              └── syntect v5.1.0
                  └── r3bl_tui v0.5.2
                      └── r3bl-cmdr v0.0.11