v0.4.0-terminal_async
[terminal_async] Remove use of TokioMutex

There are some dangers to being "cancel safe" when using async Rust.
This can be seen in the links below (to docs, and video) which outline
what the problem is.

https://docs.rs/tokio/latest/tokio/sync/struct.Mutex.html
https://www.youtube.com/watch?v=1zOd52_tUWg&t=2088s

It is better to avoid using a TokioMutex to check for cancellation
and instead to use broadcast channel for shutdown signals, just like
the code already does. The changes made in this commit are related
to removing the use of TokioMutex all together in favor of the StdMutex
since there is really no need to use it at all. And thus avoid any
potential of "cancel safe" errors cropping up!