[all] Remove needless anchor tags from lib.rs & README.md
Markdown automatically generates anchor tags for headings. The vscode
extension `Auto Markdown TOC` is used to generate the table of contents
and it was spuriously adding anchor tags to the headings.
https://marketplace.visualstudio.com/items?itemName=huntertran.auto-markdown-toc

This commit removes the anchor tags from the headings in the `lib.rs`
and (their cargo-readme generated) `README.md` files.