[core][tui] Update dep to latest unicode-width crate
This change to unicode-width 2.0.0 breaks lots of tests because it
changes the width values of certain emoji. The tests have been updated
to reflect this.

To make even more robust tests, perhaps in the future, it might be good
to store the width of characters used in variables, rather than hard
coding their values.