[all][script][log] Promote scripting.rs from rust_scratch/tls
Create 2 new crates:
- script
- log
  - support UnicodeString for output
  - update all calls to tracing::*() to work seamlessly with CustomEventFormatter
    across the entire r3bl codebase (especially r3bl_tui examples)
  - add of the `options: impl Into<TracingConfig>` pattern for configuring the
    log functionality. this is ergonomic for simple cases (for the caller) and
    provides sophisticated configuration possibilities due to its composability.

Reorganize all the existing crates as well, to ensure that
known duplicated functions are removed.

Add new glyphs that are unicode character based, not emojis,
so that the fg, bg color and attributes are applied in output.
Emojis come with their own colors, so the fg and bg colors don't
get applied.

Use Vec::with_capacity wherever possible to prevent having reallocations
which are expensive, and create a global size limit for the all the r3bl
crates.