[all] Continue memory latency performance tuning work
[ At the start ... ]

This is the 2 week long refactor that is showing performance
regression over the previous commit!

excluding the following:
- tui/md_parser/convert_to_plain_text.rs
- tui/syntax_highlighting/intermediate_types.rs
- tui/syntax_highlighting/md_parser_syn_hi/md_parser_syn_hi_impl.rs

The files in this commit had to be rolled back as part of the changes
made in the last commit were breaking all the tui tests, especially
around MD parsing. These files still need the memory optimization work:
- stack vs heap app
- use write! instead of format!

[ And then ... ]

This commit brings the performance back up to previous commit with
SHA: 5603296b484a498dafb09137b79a74eeeae9a3d3

The slow down, as far as I can tell, from running different
commits side by side, is that the computation of the UnicodeString
proved to be quite expensive to do every time. So it is beneficial
to memoize this! Running the code with flamegraph profiling didn't
really show anything useful related to this problem.

Apply heap allocation techniques to the code in the following modules
to boost performance and keep it stable (steady):
- [x] formatter.rs
- [x] editor_engine.rs
- [x] convert_to_plain_text.rs
- [x] intermediate_types.rs
- [x] md_parser_syn_hi_impl.rs