[tui] Fix sanitize_and_save_abs_pos()
There was a bug in the code dating back to when this function
was first written. This bug didn't really affect anything. However
it got exposed when the conversion to "newtype" pattern was introduced
and all the ChUnit types got replaced with Pos, Dim, Width, Height, etc.

Make variable names shorter, so that they don't have the "intended use
in the absence of concrete types" backed into the their extremely long
and unreadable names.

This is all thanks to the work related to the "newtype" pattern!