[tui] Replace use of bool with enum variants for readability
Fix bug in main event loop which was preventing the app from handling
resize events. This error was introduced in a previous refactor where
the resize event gets consumed by the main event loop, after it records
telemetry about it; instead of then also giving it to the app to
handle or process.