[core][tui] Change range.rs so that it uses CaretScrAdj type
This is more disambiguation by using proper types (now that they exist)
by favoring more specific types over more generalized types in order to
remove the need for documentation or overly long variable names (that
describe the intent, which is now captured in the type system and can
be checked by the compiler).

SelectionRange now only works with CaretScrAdj types, so there's no way
to pass a CaretRaw into the thing. And there's only one way to construct
it. And it removes the RowIndex as part of construction (by clobbering
it with an "impossible" value, rendering it clearly unusable and also
looking strange when debugging, which is the intent).

Clean up the dummy view port use case in buffer_selection_support.rs so
that it no longer passes a 0x0 viewport when it doesn't have one. And
instead passes some "impossible" value, rendering it clearly unusable
and also looking strange when debugging, which is the intent).