[all] Configure bacon jobs doc-open, check, test, etc
The default bacon doc-open job spawns a new browser window to view
the newly generated docs. And only doc-open job can be run from
the TUI using the `d` key. So this toml file changes the default
doc-open job so it does not spawn a browser window. This enables
bacon to be kept open and continusouly scanning for doc changes
and regenerating them.

It is handy to have 3 bacon instances open in split pane terminals
(in terminator, for eg) while making extensive refactoring or
rewrite changes:
1. test
2. doc-open
3. clippy or check