[core] Split RingBuffer into RingBufferHeap and RingBufferStack
Fix logic error that put both these in the same struct. The problem with
doing this is that even if a heap variant is chosen, the
RingBufferStorage enum is sized to be as large as the "stack allocated"
version, which leads to stack overflow for large N.

By splitting this struct into 2 structs, which is how it should have
been from the start (if you look at the tests, this is a dead giveaway)
this problem never arises. You choose which struct you would like to use
and it is sized appropriately (on the heap or stack).