[core][log][tui] Review clippy::large_enum_variant lint
[ color wheel ]

Allowing clippy::large_enum_variant still makes sense for the codebase.
There are situations, especially in color wheel, where to increase
memory locality, and reduce memory latency, lots of bytes are jammed
on to the stack.

[ Minor changes to smallvec size tuning ]

Based on some manual testing, tweak some size defaults for color wheel
to see if it changes the performance of the demo examples in release
mode. It didn't make too much of a difference.

[ Add more methods to RingBuffer ]

- as_slice(), push(), pop()

[ offscreen buffer ]

Remove needless enum from offscreen buffer which is what Option<T>
already does.