[core] Adjust DEFAULT_STRING_STORAGE_SIZE to 16 for performance gains
This single line of code fixes the performance regression from this
commit:
5603296b [all] Use smallvec and smallstr, memory latency, telemetry, performance tuning

If the default sizes (like DEFAULT_STRING_STORAGE_SIZE) is too high, then
it can lead to a slowdown since if more memory is allocated than what is
actually needed. So it is best to start with low values and test and tune
the software for your workloads.

Please don't rely on intuition :) Or prematurely optimize :).