[core][tui] Make Telemetry light and fast
[ telemetry.rs ]
Replace the complex implementation of telemetry with a simple one.
Instead of storing the "formatted" output of the report, store it as
a struct of numbers, which are Copy. This makes it fast to update the
report and pass it around.

Only format it when needed for display. And provide a formatter to do
this.

[ shared_global_data.rs ]

Allocate one TelemetryReportLineStorage just once. And keep it around
for the lifetime of the GlobalData struct, and simply re-use this
stack allocated buffer to "print" the "display" output of the
TelemetryHudReport into.

[ offscreen_buffer_pool.rs ]

Make sure that take() always yields a buffer. Even if the buffer is
exhausted. Also when taking back an existing buffer, make sure that
it is still valid (compare w/ current window size).