[tui] Minor optimization for converting editor content into byte array
The engine now holds a DocumentStorage struct that is re-used everytime
the document changes and needs to be parsed. This does not eliminate
the inefficiency of copying the entire document & adding CRLFs where
needed into this buffer. But it does prevent us from having to reallocate
this entire buffer, by simply clearing and reusing it.

The parser byte cache is also grown in an efficient manner, so that
small increments in reallocation are not done. Rather extensions in
capacity are done 1 page at a time, where the page size is
PARSER_BYTE_CACHE_PAGE_SIZE.