[core] Make TuiStyle and ASTStyle compatible
When using r3bl_core::ansi module functions, make it easy to supply a
TuiStyle (via new_style!) that makes it really easy to work with styles.
Then convert the TuiStyle into an array of ASTStyle (which is not as
ergonomic). This makes the API more easy to use, and leaves the
implementation of r3bl_core::ansi unchanged.

Fixes: #406