[tui] Make ReadlineAsync more robust in exit and resource cleanup
Add the ability for `ReadlineAsync` to abort the `Readline` main event loop when it is
exited (by calling the `ReadlineAsync::exit()` method). This is one of the main
differentiators between a sync blocking `read_line()` and an async non-blocking one.
Further clean up is provided that are triggered by `exit()` method to ensure that all
the tasks that spin up are cleaned up: 1. task to listen to `InputDevice`, and 2.
another task to process
`crate::readline_impl::manage_shared_writer_output::spawn_task_to_monitor_line_state_signals`.