[core] Rewrite AnsiStyledText so that it is easy to use
Make `AnsiStyledText` own the `text`, this is just a more ergonomic and
useful API than using a `&str` which introduces needless lifetimes and
lots of other unergonomic code writing, which makes this a cumbersome
API to use.

Also add lots of nice constructor functions and macros to make it a
breeze to use.