[tui] Rework spinner to eliminate lock contention
The old spinner implementation led to some strange lock contention issues
and bizarre timing problems that resulted in one task getting to the
stdout just before another one, causing the output to appear out of
the current sequence. This is very bad when printing indeterminate
progress bars! The timing and sequence matter. Intermediate ticks must
occur before the final tick!

The problems above affected stdout, stderr, and a combination of the two.
This is why OutputDevice::default() is added to default to stdout, just
for simplicity. Since it doesn't really matter if the spinner outputs
to stdout or stderr or a combination of the two (since both can be
redirected to a file). A better approach is to check for terminal
interactivity and bail if the terminal is uninteractive, rather than
assuming that stdout or stderr is better than the other.

While eliminating the lock contention and timing issues (leading to
undefined behavior) rewrite the code to be modern and use OutputDevice
and the appropriate macros.

Fixes: #424