[tui][cmdr] Overhaul TempDir and try_cd() so that their usage is safer
git.rs adds lots of new tests to make sure git commands work as intended
in the context of giti. This introduced a new scenario where cargo test
which runs tests in the same process were not working. However, cargo
nextest which runs each test as a separate process worked just fine.

The fixes to TempDir and try_cd() and all the macros that are added
including serial_async_test_with_safe_cd! ensure that all test code that
can call try_cd() has to be run serially (using serial_test). Along with
support for doing this in an async context.