[all] Update deps to latest
[Remove paste from deps for tui/Cargo.toml]

paste is no longer maintained and has been archived. It isn't really
used anywhere in the codebase, except for one declarative macro which
is defined but not used anywhere. Drop that macro.

[Update docker information and add docker-compose.yml]

Add README.md to make it easy to install and uninstall docker to verify
the `cargo build r3bl-cmdr` on a fresh machine install.

[cargo deny]

Update deny.toml to allow `ring` which is a dep of `rustls`.