WIP 12 - Revert "parse by line" strategy for parse_markdown.rs
There were 2 strategies at the start of this rewrite journey:

1. parse &[GCString] one line at a time & peek for smart list, fenced
   code block
2. create AsStrSlice and implement nom::Input on it and change all the
   existing parsers to accept AsStrSlice<'a> instead of &str

At this point, it is clear that strategy 1 is not going to work. It
requires the entire parsing code to be rewritten (needlessly).

Instead strategy 2 is the best approach, since it does not allocate
a massive DocumentStorage for each editor buffer change, and it allows
reuse of all existing parsers!

This commit removes all the Strategy 1 code.