[tui] Disable NG parser, fix flamegraph data collection, add profiling compiler profiles
NG parser is an experiment and is just too slow. O(n^2) performance is
not acceptable. More work needs to be done to address this. But we might
just deprecate this parser, since the legacy parser is really fast, and
it turns out &[GCString] -> String materialization is *REALLY* cheap,
in the order of a thousand nano sec for ~ 100K input!

Fix flamegraph arguments so that it does not have any "[unknown]"
sections, and is fast to run. Claude Code can't handle files > 256K, so
this optimization makes the flamegraph.svg file that is produced to
~70K in size.

Add cargo profiles for profiling compilation in Cargo.toml. This allows
for easy customization (along with the script_lib.nu function
run_example_with_flamegraph_profiling).