[tui] Add comprehensive snapshot testing to parser
Provide a clean way for the ex_editor example to access test
data that is used both in the example and tests (both regular
test and bench).

Add comprehensive snapshot testing tot he parser, to ensure
that it exercises and validates all the input data in the
conformance test suite (conformance_test_data folder).