[tui] Remove needless debug format! calls in each queue_render_op! call
initially the 2nd arg was meant to be a detailed debug formatted output
listing exactly what the render op is. this is just too much overhead
in the hot loop, and this debug output isn't even used if logging for
that module is disabled (which it is!)

problem remains with crossterm still using write! (flamegraph shows
this is a bottleneck)