[all] Rewrite and consolidate run.nu into a single file
Improve the bootstrap experience for this project / repo by adding
bootstrap.sh to get this project up and running on any Linux and macOS
computer.

Simplify and streamline all the project specific commands into a single
run.nu at the root level of the project.

Update all the docs so they are up to date with these changes.