[cmdr] Fix giti tests to use git binary with no signing keys
When git binary gets executed, it uses the ~/.gitconfig of the host
machine. If there is some git signing key issues there, this test will
fail. This happened to me when I was in the middle of moving around
the folder where I store my git signing keys for github.