[tui] Fix readline_async channel sizing ambiguity
This issue showed itself in the flakiness of a tests that traverses
the filesystem. The test now has a limit so the results will be more
deterministic (as we have no idea what machine the tests can be run on).

This helped shine a light on the ambiguity of what the channel size
should be. Default is contextual and the function can't assume what
channel capacity is reasonable for the caller's use case. So they have
to provide it explicitly. If None is provided, we default to the max
capacity.